Remove output on command failure
Since the `packer fmt` command does not return a non-zero exit code if
file are modified and the `packer validate` command has its own output
if a template fails validation, it does not make sense to add more to
the output of either command. This is especially true since the `packer
fmt` command will output modified files and the `packer validate`
command will mention the failing files in its output.